remove IndirectClient V1
