+++ pac4j-core/src/test/java/org/pac4j/core/client/MockIndirectClient.java
@@ -2,8 +2,6 @@ package org.pac4j.core.client;
 
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.credentials.Credentials;
-import org.pac4j.core.exception.HttpAction;
-import org.pac4j.core.logout.LogoutActionBuilder;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.redirect.RedirectAction;
 
@@ -13,15 +11,13 @@ import org.pac4j.core.redirect.RedirectAction;
  * @author Jerome Leleu
  * @since 1.9.0
  */
-public final class MockIndirectClient extends IndirectClientV1<Credentials, CommonProfile> {
+public final class MockIndirectClient extends IndirectClient<Credentials, CommonProfile> {
 
     private RedirectAction redirectAction;
 
     private ReturnCredentials returnCredentials;
 
     private CommonProfile profile;
-
-    private LogoutActionBuilder<CommonProfile> logoutActionBuilder;
 
     public MockIndirectClient(final String name) {
         setName(name);
@@ -39,30 +35,12 @@ public final class MockIndirectClient extends IndirectClientV1<Credentials, Comm
     }
 
     @Override
-    protected RedirectAction retrieveRedirectAction(final WebContext context) throws HttpAction {
-        return redirectAction;
-    }
+    protected void internalInit(final WebContext context) {
+        super.internalInit(context);
 
-    @Override
-    protected Credentials retrieveCredentials(final WebContext context) throws HttpAction {
-        return returnCredentials.get();
-    }
-
-    @Override
-    protected CommonProfile retrieveUserProfile(final Credentials credentials, final WebContext context) throws HttpAction {
-        return profile;
-    }
-
-	@Override
-	protected RedirectAction retrieveLogoutRedirectAction(final WebContext context, final CommonProfile currentProfile, final String targetUrl) {
-		return logoutActionBuilder.getLogoutAction(context, currentProfile, targetUrl);
-	}
-
-    public LogoutActionBuilder<CommonProfile> getLogoutActionBuilder() {
-        return logoutActionBuilder;
-    }
-
-    public void setLogoutActionBuilder(final LogoutActionBuilder<CommonProfile> logoutActionBuilder) {
-        this.logoutActionBuilder = logoutActionBuilder;
+        setRedirectActionBuilder(ctx -> redirectAction);
+        setCredentialsExtractor(ctx -> returnCredentials.get());
+        setAuthenticator((cred, ctx) -> cred.setUserProfile(profile));
+        setLogoutActionBuilder(getLogoutActionBuilder());
     }
 }
