+++ pac4j-oauth/src/test/java/org/pac4j/oauth/run/RunBitbucketClient.java
@@ -1,6 +1,6 @@
 package org.pac4j.oauth.run;
 
-import org.pac4j.core.client.IndirectClientV1;
+import org.pac4j.core.client.IndirectClient;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.profile.Gender;
 import org.pac4j.core.profile.ProfileHelper;
@@ -34,7 +34,7 @@ public class RunBitbucketClient extends RunClient {
     }
 
     @Override
-    protected IndirectClientV1 getClient() {
+    protected IndirectClient getClient() {
         BitbucketClient client = new BitbucketClient();
         client.setKey("bjEt8BMpLwFDqZUvp6");
         client.setSecret("NN6fVXRTcV2qYVejVLZqxBRqHgn3ygD4");
