+++ pac4j-oauth/src/test/java/org/pac4j/oauth/run/RunLinkedIn2Client.java
@@ -1,6 +1,6 @@
 package org.pac4j.oauth.run;
 
-import org.pac4j.core.client.IndirectClientV1;
+import org.pac4j.core.client.IndirectClient;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.run.RunClient;
 import org.pac4j.core.profile.Gender;
@@ -36,7 +36,7 @@ public final class RunLinkedIn2Client extends RunClient {
     }
 
     @Override
-    protected IndirectClientV1 getClient() {
+    protected IndirectClient getClient() {
         final LinkedIn2Client client = new LinkedIn2Client();
         client.setKey("gsqj8dn56ayn");
         client.setSecret("kUFAZ2oYvwMQ6HFl");
