+++ pac4j-oauth/src/test/java/org/pac4j/oauth/run/RunOkClient.java
@@ -1,6 +1,6 @@
 package org.pac4j.oauth.run;
 
-import org.pac4j.core.client.IndirectClientV1;
+import org.pac4j.core.client.IndirectClient;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.run.RunClient;
 import org.pac4j.core.profile.Gender;
@@ -62,7 +62,7 @@ public final class RunOkClient extends RunClient {
     }
 
     @Override
-    protected IndirectClientV1 getClient() {
+    protected IndirectClient getClient() {
         final OkClient okClient = new OkClient();
         okClient.setKey("1139019264");
         okClient.setPublicKey("CBAPAFOEEBABABABA");
